Move code modules into separate files.
