+++ keywords.txt
@@ -1,4 +1,3 @@
 DcsBios	KEYWORD1
 processChar	KEYWORD2
-onDcsBiosWrite	KEYWORD3
 sendDcsBiosMessage KEYWORD3
